refactor cursor
