+++ src/app/home/landing.css
@@ -1,14 +1,10 @@
-html {
-  cursor: none;
-  --text: #fff;
-  scroll-behavior: smooth;
-}
-
 section {
   height: 100vh;
   display: flex;
+  flex-direction: column;
   justify-content: center;
   align-items: center;
+  gap: 4rem;
 }
 
 h2 {
@@ -48,54 +44,10 @@ h2 {
   background-repeat: no-repeat;
   background-size: cover;
 }
-
-.cursor {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 2.5rem;
-  height: 2.5rem;
-  z-index: 2;
-  pointer-events: none;
-}
-
-.cursor div {
-  position: absolute;
-  display: grid;
-  place-items: center;
-}
-
-.cursor div div {
-  border: 1px solid #fff;
-  border-radius: 50%;
-  animation: pulse 2.5s linear infinite;
-  box-shadow: 0 0 50px 5px #d7abff6e;
-}
-
-.cursor div:nth-child(1),
-.cursor div:nth-child(2) {
-  width: 100%;
-  height: 100%;
-}
-
-.cursor div:nth-child(1) {
-  transition: transform 0.2s ease-out;
-}
-
-.cursor div:nth-child(2) {
-  transition: transform 0.1s ease-out;
-}
-
-.cursor div:nth-child(2) div {
-  background: #fff;
-  border-radius: 50%;
-  width: 4px;
-  height: 4px;
-}
-
 *[data-animate="true"] {
   opacity: 0;
 }
+
 
 .animate-slide-down {
   animation: slide-down 300ms ease-out forwards;
